+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 16Apr2019
 - Added synapse location to `util\network\getLinkedNeurons.m` and edited the output parsing. If there's only a single output, all arguments will be returned as a table.
 - Updated functions using `util\network\getLinkedNeurons.m` to be compatible with new output
+- New 2D synapse function `util\render\synapseMarker.m`
 
 ### 14Apr2019
 - Added function for returning renders in a figure (`util\renders\renderWhos.m`)